==> ML in business/09 integration/Step_2_PREDICT.ipynb <==
### Проверка работоспособности и качества пайплайна

Здесь мы еще не запускаем никакое API, а загружаем модель (pipeline) напрямую и проверяем на отложенной (тестовой) выборке

import numpy as np
import pandas as pd
from sklearn.metrics import precision_recall_curve
import dill

import warnings
warnings.filterwarnings('ignore')

dill._dill._reverse_typemap['ClassType'] = type

X_test = pd.read_csv("X_test.csv")
y_test = pd.read_csv("y_test.csv")

X_test.head(3)

with open('logreg_pipeline.dill', 'rb') as in_strm:
    pipeline = dill.load(in_strm)

pipeline

preds = pipeline.predict_proba(X_test)[:,1]
pred_df = pd.DataFrame({'preds':preds})
pred_df.to_csv('test_predictions.csv', index=None)

preds[:10]

precision, recall, thresholds = precision_recall_curve(y_test, preds)
fscore = (2*precision*recall)/(precision + recall)
ix = np.argmax(fscore)
print('Best Threshold=%f, F-Score=%.3f, Precision=%.3f, Recall=%.3f' % (thresholds[ix], 
                                                                        fscore[ix],
                                                                        precision[ix],
                                                                        recall[ix]))

---
