==> ML in business/09 integration/Step_3_FLASK.ipynb <==
### При внедрении

* Определить формат json 
* Определить ip-адрес и порт, на который будут поступать данные
* Создать во Flask необходимые роутеры:
```
@app.route('/predict_example', method='POST')
def predict_example():
```
* Перенести во Flask все функции преобразования данных
    * формат данные, приходящих от фронт-системы, может отличаться от формата исторических данных, использовавшихся при построении модели; в результате данные на вход модели должны поступить ровно в том виде, в каком было обучена модель.
* Загрузить обученные модели
* Настроить логгирование

### Flask

from flask import Flask, request, jsonify
import pandas as pd
import dill

Пробный запуск Flask

app = Flask(__name__)
# run_with_ngrok(app) # это говно только в колабе нужно

@app.route('/a')
def hello():
    return "Hello world!"

if __name__ == '__main__':
    app.run()

Заходим по ссылке, вписываем адрес - видим Hello world!

### Создаем сервис для обработки запросов к модели

from flask import Flask, request, jsonify
import pandas as pd
import dill

with open('logreg_pipeline.dill', 'rb') as in_strm:
    model = dill.load(in_strm)

X_test = pd.read_csv("X_test.csv")
y_test = pd.read_csv("y_test.csv")

Запустить сервис и не глушить его, пока работаем

# Обработчик и запуск Flask
app = Flask(__name__)

@app.route("/", methods=["GET"])
def general():
    return "Welcome to prediction process"

@app.route("/predict", methods=["POST"])
def predict():
    data = {"success" : False}
    
    # ensure an image was properly uploaded to our endpoint
    description, company_profile, benefits = "", "", ""
    request_json = request.get_json()
    
    if request_json['description']:
        description = request_json['description']
        
    if request_json['company_profile']:
        company_profile = request_json['company_profile']
        
    if request_json['benefits']:
        benefits = request_json['benefits']
    
    print(description)
    preds = model.predict_proba(pd.DataFrame({
        'description':[description],
        'company_profile':[company_profile],
        'benefits':[benefits]
    }))
    
    data['predictions'] = preds[:,1][0]
    data['description'] = description
    # indicate that the request was a success
    data['success'] = True
    print('OK')
    
    return jsonify(data)

if __name__ == '__main__':
    app.run()

---
